% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.file.metadata}
\alias{create.file.metadata}
\title{Retrieve metadata about NetCDF-format files.}
\usage{
create.file.metadata(f, variable.name.map)
}
\arguments{
\item{f}{The list of NetCDF files.}

\item{variable.name.map}{A named character vector mapping standard variable names (tmax, tmin, prec) to NetCDF variable names.}
}
\value{
A list containing time data (ts), dimension sizes (dim.size), dimension axes (dim.axes), source units (src.units), destination units (dest.units), a mapping from variables to files (v.f.idx), and a projection, if available.
}
\description{
Retrieve metadata about NetCDF-format files.
}
\details{
Given a list of NetCDF files and a mapping from standard variable names (tmax, tmin, prec) to NetCDF variable names, retrieve a set of standardized metadata.
}
\examples{
\dontrun{
## Get metadata about a single input file.
input.files <- c("pr_NAM44_CanRCM4_ERAINT_r1i1p1_1989-2009.nc")
f <- lapply(input.files, ncdf4::nc_open)
f.meta <- create.file.metadata(f, variable.name.map)
}
}

