% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.climdex.functions}
\alias{get.climdex.functions}
\title{Returns a list of Climdex functions, with parameters curried in.}
\usage{
get.climdex.functions(vars.list, fclimdex.compatible = TRUE)
}
\arguments{
\item{vars.list}{The variable list, as created by \code{\link{get.climdex.variable.list}}.}

\item{fclimdex.compatible}{Whether to create fclimdex compatible functions.}
}
\value{
A list of functions, named by the variable they compute.
}
\description{
Returns a list of Climdex functions, with parameters curried in.
}
\details{
This function takes a variable list (as created by \code{\link{get.climdex.variable.list}}) and creates a list of functions corresponding to the specified indices, with parameters such as time resolution curried in. This allows for these functions to be called with just the \code{climdexInput} object as an argument, easing the automation of computing indices.
}
\examples{
## Get Climdex functions for a variable list with all appropriate params
## curried in, so that all they take is a ClimdexInput object.
cdx.funcs <- get.climdex.functions(get.climdex.variable.list(c("tmax", "tmin")))
}

