% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.data}
\alias{get.data}
\title{Retrieve and convert data to correct units and dimensions.}
\usage{
get.data(f, v, subset, src.units, dest.units, dim.axes)
}
\arguments{
\item{f}{The NetCDF file to retrieve data from; an object of class \code{ncdf4}.}

\item{v}{The variable to retrieve data from.}

\item{subset}{The subset to retrieve.}

\item{src.units}{The source units to convert data from.}

\item{dest.units}{The destination units to convert to.}

\item{dim.axes}{The dimension axes to be used.}
}
\value{
The retrieved and massaged data.
}
\description{
Retrieve and convert data to correct units and dimensions.
}
\details{
This function retrieves NetCDF data for the specified subset from the specified file and variable; converts from \code{src.units} to \code{dest.units}, transposes the data to (T, S) dimensionality, and returns the result.
}
\examples{
\donttest{get.data(f, "pr", list(Y=3), "kg m-2 s-1", "kg m-2 s-1", c(X="lon",Y="lat",T="time"))}
}

