% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.thresholds.metadata}
\alias{get.thresholds.metadata}
\title{Retrieve threshold metadata}
\usage{
get.thresholds.metadata(var.names)
}
\arguments{
\item{var.names}{A vector containing names of available variables (tmax, tmin, prec).}
}
\value{
A list containing metadata for each of the six thresholds.
}
\description{
Retrieve threshold metadata
}
\details{
Returns units, long names, locations within the climdexInput data structure, and whether time data should be included given the variable information available.
}
\examples{
thresholds.meta <- get.thresholds.metadata("prec")
}

