\name{get.var.file.idx}
\alias{get.var.file.idx}
\title{Create mapping from variables to files.}
\usage{
get.var.file.idx(variable.name.map, v.list)
}
\arguments{
  \item{variable.name.map}{A mapping from standardized
  names (tmax, tmin, prec) to NetCDF variable names.}

  \item{v.list}{A list containing a vector of variables in
  each file.}
}
\description{
Create mapping from variables to files.
}
\details{
Given a variable name map and list of variables in each
file, determine a mapping from variables to files.
}
\examples{
## FIXME
}

