% Generated by roxygen2 (4.0.2): do not edit by hand
\name{thresholds.close}
\alias{thresholds.close}
\title{Close thresholds file(s)}
\usage{
thresholds.close(thresholds.nc)
}
\arguments{
\item{thresholds.nc}{A list of objects of class \code{ncdf4}, or NULL}
}
\description{
Close thresholds file(s)
}
\details{
This function closes one or more thresholds files.
}
\examples{
\dontrun{
## Open a single thresholds file, then close it.
thresholds.files <- c("thresh.nc")
thresh <- thresholds.open(thresholds.files)
thresholds.close(thresh)
}
}

