% Generated by roxygen2 (4.0.2): do not edit by hand
\name{thresholds.open}
\alias{thresholds.open}
\title{Open thresholds file(s)}
\usage{
thresholds.open(thresholds.files)
}
\arguments{
\item{thresholds.files}{A character vector containing the names of thresholds files.}
}
\value{
A list of objects of class \code{ncdf4}, or NULL if thresholds.files is NULL.
}
\description{
Open thresholds file(s)
}
\details{
This function opens one or more thresholds files and returns the \code{ncdf4} objects as a list.
}
\examples{
\dontrun{
## Open a single thresholds file
thresholds.files <- c("thresh.nc")
thresh <- thresholds.open(thresholds.files)
}
}

