\name{write.climdex.results}
\alias{write.climdex.results}
\title{Write out computed climdex results}
\usage{
write.climdex.results(climdex.results, chunk.subset, cdx.ncfile, dim.size)
}
\arguments{
  \item{climdex.results}{The results to write out.}

  \item{chunk.subset}{The corresponding subset.}

  \item{cdx.ncfile}{The list of NetCDF files to write the
  results out to.}

  \item{dim.size}{The overall size of the input data.}
}
\description{
Write out computed climdex results
}
\details{
Given a set of Climdex results, a subset, a set of files,
and dimension sizes, write out the data to the appropriate
files.
}
\examples{
## FIXME
}

