% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.climdex.cmip5.filenames}
\alias{create.climdex.cmip5.filenames}
\title{Creates a list of CMIP5-compliant filenames reflecting the input data.}
\usage{
create.climdex.cmip5.filenames(fn.split, vars.list)
}
\arguments{
\item{fn.split}{A vector containing named components, as created by \code{get.split.filename.cmip5}.}

\item{vars.list}{A vector containing names of variables, as created by \code{\link{get.climdex.variable.list}}.}
}
\value{
A vector containing filenames corresponding to the variables and filename bits supplied.
}
\description{
Creates a list of CMIP5-compliant filenames reflecting the input data.
}
\details{
This function takes a split filename (as created by \code{get.split.filename.cmip5}) and a list of variables and creates corresponding filenames for the given variables.
}
\examples{
\dontrun{
library(ncdf4.helpers)
## Split out filename bits for use below...
fn <- "pr_day_BCCAQ+ANUSPLIN300+MRI-CGCM3_historical+rcp85_r1i1p1_19500101-21001231.nc"
fn.split <- get.split.filename.cmip5(fn)

## Create filenames with time data and variable appropriately replaced.
filenames <- create.climdex.cmip5.filenames(fn.split, c("rx5dayETCCDI_mon", "tn90pETCCDI_yr"))
}
}

