% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.climdex.variable.list}
\alias{get.climdex.variable.list}
\title{Returns a list of Climdex variables given constraints}
\usage{
get.climdex.variable.list(source.data.present, time.resolution = c("all",
  "annual", "monthly"), climdex.vars.subset = NULL)
}
\arguments{
\item{source.data.present}{A vector of strings naming the data that's present; at least one of (tmin, tmax, prec, tavg).}

\item{time.resolution}{The time resolutions to compute indices at. See \code{\link{create.indices.from.files}}.}

\item{climdex.vars.subset}{A character vector of lower-case names of Climdex indices to calculate (eg: tr, fd, rx5day). See \code{\link{create.indices.from.files}}.}
}
\value{
A character vector containing variable names with time resolutions appended.
}
\description{
Returns a list of Climdex variables given constraints.
}
\details{
This function takes a character vector which specifies what source data is present and a time resolution, and generates a list of names consisting of the variable and the time resolution, separated by an underscore.
}
\examples{
## Get all variables which require tmin and/or tmax, for all time resolutions.
var.list1 <- get.climdex.variable.list(c("tmax", "tmin"))

## Get all variables which require prec with an annual time resolution.
var.list2 <- get.climdex.variable.list("prec", time.resolution="annual")

## Get the intersection of a set list of vars and available data.
sub.vars <- c("su", "id", "tr", "fd", "gsl", "csdi", "wsdi", "r10mm")
var.list3 <- get.climdex.variable.list("tmax", climdex.vars.subset=sub.vars)
}
\seealso{
\code{\link{create.indices.from.files}}
}

