% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.var.file.idx}
\alias{get.var.file.idx}
\title{Create mapping from variables to files.}
\usage{
get.var.file.idx(variable.name.map, v.list)
}
\arguments{
\item{variable.name.map}{A mapping from standardized names (tmax, tmin, prec) to NetCDF variable names.}

\item{v.list}{A list containing a vector of variables in each file.}
}
\value{
A vector mapping standardized variable names (tmax, tmin, prec) to indices in the file list.
}
\description{
Create mapping from variables to files.
}
\details{
Given a variable name map and list of variables in each file, determine a mapping from variables to files.
}
\examples{
\dontrun{
## Get mapping for a single file.
input.files <- c("pr_NAM44_CanRCM4_ERAINT_r1i1p1_1989-2009.nc")
f <- lapply(input.files, ncdf4::nc_open)
v.list <- lapply(f, ncdf4.helpers::nc.get.variable.list, min.dims=2)
v.f.idx <- get.var.file.idx(variable.name.map, v.list)
}
}

